PARQUET-1687: Update release process
Update the link of the keys file to the official one. The official link
is required by the Apache release process.
Update the usage of the prepare script to use the RC number as a
parameter so they can create RC tags.
Add usage of the new finalize script that creates the final release tag
and sets the new development version.
As we are using RC tags and add the final release tag only after the
vote passes we do not need to do anything in case of the vote fails.