Clearning out all contents of parquet staging
We are clearing out all contents of the parquet staging directory to
deploy a new website, following the accepted Apache best practices. In
these best practices, the `asf-site` and the `asf-staging` branches host
the generated website content for publication.

The code that generates these templates lives on a separate branch.